+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+### Fixed
 - Updated DEVELOPMENT.md to reflect correct file structure
+
+### Added
+- Added sub not_today_enabled_configs for the updates report
 
 ## [0.9.16] - 2026-08-28